+++ src/invoices/utils.js
@@ -1,7 +1,15 @@
-export const getAdjustmentFromPreset = ({ description, prorate, relationToTotal, type, defaultAmount }) => ({
+export const getAdjustmentFromPreset = ({
+  description,
+  prorate,
+  relationToTotal,
+  type,
+  defaultAmount,
+  exportToAccounting,
+}) => ({
   description,
   prorate,
   relationToTotal,
   type,
   value: defaultAmount,
+  exportToAccounting,
 });